test(integration): move real-LLM OrchestratorTool tests to test_agent_integration.py and mark as @pytest.mark.slow; keep unit tests in test_agents.py
