+++ src/coding_assistant/agents/tests/test_agents.py
@@ -1,13 +1,10 @@
-from pathlib import Path
-from unittest.mock import AsyncMock, MagicMock, patch
-
 import pytest
 
 from coding_assistant.agents.callbacks import NullCallbacks
 from coding_assistant.agents.execution import handle_tool_call
-from coding_assistant.agents.types import Agent, AgentOutput, TextResult, Tool
+from coding_assistant.agents.types import Agent, TextResult, Tool
 from coding_assistant.config import Config
-from coding_assistant.tools.tools import FeedbackTool, OrchestratorTool
+from coding_assistant.tools.tools import FeedbackTool
 
 TEST_MODEL = "openai/gpt-5-mini"
 
@@ -82,117 +79,6 @@ async def test_feedback_tool_after_feedback():
         }
     )
     assert result.content == "Ok"
-
-
-@pytest.mark.long
-@pytest.mark.asyncio
-async def test_orchestrator_tool():
-    config = create_test_config()
-    tool = OrchestratorTool(config=config)
-    result = await tool.execute(parameters={"task": "Say 'Hello, World!'"})
-    assert result.content == "Hello, World!"
-
-
-@pytest.mark.long
-@pytest.mark.asyncio
-async def test_orchestrator_tool_resume():
-    config = create_test_config()
-    first = OrchestratorTool(config=config)
-
-    result = await first.execute(parameters={"task": "Say 'Hello, World!'"})
-    assert result.content == "Hello, World!"
-
-    second = OrchestratorTool(config=config, history=first.history)
-    result = await second.execute(
-        parameters={"task": "Re-do your previous task, just translate your output to German."}
-    )
-    assert result.content == "Hallo, Welt!"
-
-
-@pytest.mark.long
-@pytest.mark.asyncio
-async def test_orchestrator_tool_instructions():
-    config = create_test_config()
-    tool = OrchestratorTool(config=config)
-    result = await tool.execute(
-        parameters={
-            "task": "Say 'Hello, World!'",
-            "instructions": "When you are told to say 'Hello', actually say 'Servus', do not specifically mention that you have replaced 'Hello' with 'Servus'.",
-        }
-    )
-    assert result.content == "Servus, World!"
-
-
-def _create_confirmation_orchestrator():
-    config = create_test_config()
-    config.shell_confirmation_patterns = ["^echo"]
-    tool = OrchestratorTool(config=config)
-    parameters = {
-        "task": "Execute shell command 'echo Hello World' and verbatim output stdout. If the command execution is denied, output 'Command execution denied.'",
-    }
-    return tool, parameters
-
-
-@pytest.mark.asyncio
-async def test_shell_confirmation_positive():
-    tool, parameters = _create_confirmation_orchestrator()
-
-    with patch("coding_assistant.tools.tools.create_confirm_session") as mock_create_confirm:
-        mock_session = AsyncMock()
-        mock_session.prompt_async.return_value = True
-        mock_create_confirm.return_value = mock_session
-
-        result = await tool.execute(parameters=parameters)
-        assert result.content.strip() == "Hello World"
-
-
-@pytest.mark.asyncio
-async def test_shell_confirmation_negative():
-    tool, parameters = _create_confirmation_orchestrator()
-
-    with patch("coding_assistant.tools.tools.create_confirm_session") as mock_create_confirm:
-        mock_session = AsyncMock()
-        mock_session.prompt_async.return_value = False
-        mock_create_confirm.return_value = mock_session
-
-        result = await tool.execute(parameters=parameters)
-        assert result.content.strip() == "Command execution denied."
-
-
-def _create_tool_confirmation_orchestrator():
-    config = create_test_config()
-    config.tool_confirmation_patterns = ["^execute_shell_command"]
-    tool = OrchestratorTool(config=config)
-    parameters = {
-        "task": "Use the execute_shell_command to echo 'Hello, World!' and verbatim output stdout. If the tool execution is denied, output 'Tool execution denied.'",
-    }
-    return tool, parameters
-
-
-@pytest.mark.asyncio
-async def test_tool_confirmation_positive():
-    tool, parameters = _create_tool_confirmation_orchestrator()
-
-    with patch("coding_assistant.agents.execution.create_confirm_session") as mock_create_confirm:
-        mock_session = AsyncMock()
-        mock_session.prompt_async.return_value = True
-        mock_create_confirm.return_value = mock_session
-
-        result = await tool.execute(parameters=parameters)
-        assert result.content.strip() == "Hello, World!"
-
-
-@pytest.mark.asyncio
-async def test_tool_confirmation_negative():
-    tool, parameters = _create_tool_confirmation_orchestrator()
-
-    with patch("coding_assistant.agents.execution.create_confirm_session") as mock_create_confirm:
-        mock_session = AsyncMock()
-        mock_session.prompt_async.return_value = False
-        mock_create_confirm.return_value = mock_session
-
-        result = await tool.execute(parameters=parameters)
-        assert result.content.strip() == "Tool execution denied."
 
 
 class FakeBigOutputTool(Tool):
